Data.h ----------------------- added the stub for a generic = operator overload to allow filterManager to copy Data objects without having to type case and explicitly calling the specific operator for ImgData or FPGAData
FilterManager.h
-----------------------
Removed getIndexByID func. idea of the IDHasher ADT is so that we can
search by keys and not have to use indices. Getting index to a specific
key is therefore useless since we would prefer to use keys instead.
Potential discussion is where if user wants to delete specifically the
nth key in the filterChain. Currently the best solution is for user to
obtain a list of the IDs in filterChain, look up index n, and use the
key to complete the deletion. 

FilterManager.cpp
--------------------------
Organized to follow header file.
variable names updated.
Core implementations are full versions of insertFilter, and
deleteFilter. These two functions should be tested heavily. Design
decision here is that if these two functions work, other overloaded
functions may be generalized to these two cases with ease and testing
can be done quicker.


NO LOGGING YET