fixed and mem tested rgbfilter
changed filtration implementation. use unsigned char now instead of char
after consulting opencv documentation. SetValue function altered to
reflect this change. included math lib to use the max, min and abs
functions to simplify code.

in the rgbfilterTEST, found errors in FilterFactory... mem allocation
error... not sure what causes it since if the rgbfilter constructor is
used explicitly, no error arises. go into the rgbfilterTEST cpp to
uncomment the T_constructor unit to reproduce the result. I don't think
this is a critical error (so as long as we do not use filterfactory) but
it is still worthy of fixing before Friday.